refactor: 機能単位のディレクトリ構成に整理
