+++ package.json
@@ -29,7 +29,7 @@
   "devDependencies": {
     "@eslint/js": "^9.21.0",
     "@tailwindcss/vite": "^4.1.18",
-    "@types/node": "^22.14.0",
+    "@types/node": "^22.19.3",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
     "@types/three": "^0.182.0",
