+++ src/app/components/index.tsx
@@ -1,0 +1,3 @@
+export { default as MenuOverlay } from './MenuOverlay';
+export { default as SectionIndicator } from './SectionIndicator';
+export { default as TopNav } from './TopNav';
