+++ src/features/about/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './components/AboutSection';
