+++ src/features/contact/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './components/ContactSection';
