+++ src/features/hero/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './components/HeroSection';
