+++ src/shared/components/DetailModal.tsx
