+++ src/shared/components/FloatingPanel.tsx
