+++ src/shared/constants/sections.ts
