+++ src/shared/data/translations.ts
